polish code structure, comments and add gitignore
