+++ .gitignore
@@ -1,0 +1,59 @@
+# Compiled files
+*.py[cod]
+*.a
+*.o
+*.so
+*.pyd
+__pycache__
+
+# Ignore .c files by default to avoid including generated code. If you want to
+# add a non-generated .c extension, use `git add -f filename.c`.
+*.c
+
+# Other generated files
+MANIFEST
+astropy/version.py
+astropy/cython_version.py
+astropy/wcs/include/wcsconfig.h
+astropy/_erfa/core.py
+astropy/_erfa/core.pyx
+
+# Sphinx
+_build
+_generated
+docs/api
+
+
+# Packages/installer info
+*.egg
+*.egg-info
+dist
+build
+eggs
+.eggs
+parts
+bin
+var
+sdist
+develop-eggs
+.installed.cfg
+distribute-*.tar.gz
+
+# Other
+.cache
+.tox
+.*.swp
+.*.swo
+*~
+.project
+.pydevproject
+.settings
+.coverage
+cover
+htmlcov
+
+# Mac OSX
+.DS_Store
+
+# PyCharm
+.idea
