+++ run_manager2.py
@@ -3,76 +3,18 @@ import os.path
 import glob
 from fnmatch import fnmatch
 import time
-import datetime
 import sys
 import shutil
 import re
 import numpy as np
 import signal
+from .simulation_instances import SimulationInstance
 
 
-"""
-Global configurations
-#TODO: move the configuration to a text file called 'SiMon.conf'. Parse the config file with regex.
-"""
-sim_dir = '/Users/maxwell/Works/nbody6/Ncode/run'
-
-"""
-A simulation instance is a single simulation task which the user requests to finish.
-It is associated with 1) a set of initial conditions specified in the input file,
-2) a (bash) script to start up the code, 3) the status of the simulation (RUN/STOP/model time,
-start timestamp, last output timestamp, parent simulation ID if it is a restart, etc),
-and 4) the ending time of the simulation.
-"""
-class SimulationInstance(object):
-    def __init__(self, id, name, fulldir, status, t_min = 0, t_max = 0, restarts = None):
-        self.id = id
-        self.name = name # usually the name of the simulation directory
-        self.fulldir = fulldir # the full path of the simulation directory
-        self.status = status # RUN, STOP, RESTARTED
-        self.errortype = ''
-        self.t_min = t_min
-        self.t_max = t_max
-        self.t_max_extended = t_max # extended t_max by restarts
-        self.mtime = 0
-        self.ctime = 0
-        self.cid = -1 # the candidate instance ID to restart in case crashes
-                      # (-1: no candidate, restart from itself;)
-                      # (>0: restart from the candidate. If the candidate cannot restart, try siblings)
-        self.level = 0
-        self.parent_id = -1
-        if restarts == None:
-            self.restarts = list()
-        else:
-            self.restarts = restarts # children
-
-    """
-    Traverse the directory structure, and print the structure information in the terminal.
-
-    A hierarchical directory structure may form for a simulation that has been started for multiple times.
-    For instance, a simulation is running on the directory '/sim1'. It crashes at T=120. So SiMon
-    restarts it by creating a restart directory '/sim1/restart1'. 'restart1' runs until T=200, and then
-    again crashes. So SiMon creates '/sim1/restart1/restart1' in attempt to start from T=200.
-    """
-    def __repr__(self, level=0):
-        placeholder_dash = "|---"+'-'*(level*4)
-        placeholder_space = "    "+' '*(level*4)
-        ctime_str = datetime.datetime.fromtimestamp(self.ctime).strftime('%Y-%m-%d %H:%M:%S')
-        mtime_str = datetime.datetime.fromtimestamp(self.mtime).strftime('%Y-%m-%d %H:%M:%S')
-        info = "%s\t%s\t%s\n%s%s\tT=[%d-%d]\t%s\tCID=%d\tlevel=%d" % (repr(self.name), ctime_str, mtime_str,
-                placeholder_space, self.status, self.t_min, self.t_max, self.errortype,self.cid,self.level)
-        ret = "%d%s%s\n" % (self.id, placeholder_dash, info)
-        #ret = "    "*level+str(self.id)+repr(self.name)+"\n"
-        for child in self.restarts:
-            ret += child.__repr__(level+1)
-        return ret
+# TODO: move the configuration to a text file called 'SiMon.conf'. Parse the config file with regex.
+sim_dir = '/Users/penny/Works/simon_project/nbody6/Ncode/run'  # Global configurations
 
 
-
-
-"""
-The Implementation of SiMon.
-"""
 class Run_Manager():
 
     def __init__(self, pidfile=None, stdin='/dev/tty', stdout='/dev/tty', stderr='/dev/tty',
@@ -95,10 +37,6 @@ class Run_Manager():
         self.tcrit = 100
         os.chdir(cwd)
 
-
-
-
-
     """
     Prompt to the user to input the simulation ID.
     """
@@ -109,7 +47,6 @@ class Run_Manager():
             if raw_input('Your input is \n\t'+str(ids)+', confirm? [Y/N] ').lower() == 'y':
                 confirmed = True
                 return ids
-
 
     """
     Traverse the simulation file structure tree, until the leaf (i.e. no restart directory) or
